refactor(maestro): removed sb shrothand
