+++ packages/maestro/package.json
@@ -16,9 +16,9 @@
       "import": "./lib/vite.js",
       "require": "./lib/vite.js"
     },
-    "./sb": {
-      "import": "./lib/sb.js",
-      "require": "./lib/sb.js"
+    "./storybook": {
+      "import": "./lib/storybook.js",
+      "require": "./lib/storybook.js"
     }
   },
   "scripts": {
